Split Case feature specs into two files
Rather than having two vaguely related groups of feature tests (in that
they both test Cases) within one file, split these out into different
files based on the page they are testing.
